compute hashes to keys for ResolutionSharedMemory
Summary: The test changes ended up being a bit more involved than hoped - turns out that `assert_response` used to clear all handles, so we weren't computing hashes to keys for any of the modules keyed by handle. The test has been complicated because of that; we also test all the AST modules as well now.

Reviewed By: grievejia

Differential Revision: D14716514

fbshipit-source-id: 376fdbadd30cc286d9bf55de021fbb5e8ff5eb1b