compute hashes to keys for Protocols
Summary:
Oversaw a different set of protocols for servers.

Note that this also implies that we were never invalidating inherited protocols in the server :(

Also makes the protocol take a IntKey of 0 instead of a random string key - the 0 of int is more intuitive as a "Placeholder key that all callsites respect" over a random name, to me at least.

Reviewed By: dkgi

Differential Revision: D14717530

fbshipit-source-id: f59809bb368a48badb593187f166cdb513e2df78