creating TypeConstraints module
Summary:
Supporting less or equal solving with variables on both sides requires explicit support for the distinction between lower and upper bounds on variables.
Our existing implementation conflated the two inconsistently.

This version maintains an interval for each variable, `lower_bound` ≤  variable ≤ `upper_bound`.
You can add a lower bound or upper bound to a variable, which is implemented based on this implication:
   `lower_bound_1 join lower_bound_2` ≤  variable ≤ `upper_bound_1 meet upper_bound_2` =>
   `lower_bound_1` ≤  variable ≤ `upper_bound_1` && `lower_bound_2` ≤  variable ≤ `upper_bound_2`

After adding a series of upper and lower bounds, you can then extract a "solution" from the constraints which selects the lowest non-bottom value it can for the variables such that all of the given constraints are followed, along with the constraints implied by the type of the variable (i.e. explicit or upper bounded).

Circular dependencies in the constraint set lead to no solution being returned.

The last element of the interface is a partial solution extraction, which attempts to generate a solution for the subset of variables requested, returning that partial solution along with the remaining constraints with the constraints substituted according to the partial solution.
This allows us to remove the need for simulated bound variables in the handling of generic callables in solve_less_or_equal.

One could imagine defining another operation, an exclusive partial solution extraction, where you generate a sub solution for all variables not-specified, yielding that subsolution and the specified variables constraints substituted with that subsolution.
This could eliminate the other use of simulated bound variables, generic protocols.  To keep this simpler, I'm deferring that work, as it requires a little bit more work on the protocol side to take advantage of it.

I wanted to keep this outside of TypeOrder since this is conceptually separate, and TypeOrder is getting super big.
This distinction also allowed me to do much purer unit tests on this module, which is nice given how confusing this can be even on its own.

Reviewed By: dkgi

Differential Revision: D14672430

fbshipit-source-id: 993462d34bfd373a8b33d49ba666fd53300e4d9c