Taking advantage of improved TypeConstraints to implement bidirectional solve_less_or_equal
Summary:
With the benefit of TypeConstraints we can (relatively) straightforwardly retrofit solve_less_or_equal to handle variables on both sides of the comparison.
This has many implications all over the codebase, and also allows us to remove a bunch of hacks, as well as correctly implement callable solving as well as correctly handling parametric variance.
Interval constraints are very fun :)

Note1: this diff is big, but it's smaller than it looks because it's currently counting all of the indenting I needed to do in order to put much of TypeOrder into a module in order to get the recursive dependency on TypeConstraints to work.  There is probably some splitting of this diff that could be done but it would be pretty tricky, and this is all unified under the implications of allowing bidirectional solves.  Still happy to give splitting a go if this is just too much to review.
Note2: this diff tightened up solution to a degree where the fact that we're only considering one overload matches started adding failures to calls to map.  This is fixed in the follow up diff.

Reviewed By: dkgi

Differential Revision: D14658240

fbshipit-source-id: ae9a2cf066bbab14eda7a35a5add3c0f8bd5f88c