purge protocols per handle
Summary:
We had a bug where we were never removing elements from the master list of protocols. This meant that if you registered a class as a protocol and then removed the protocol superclass, Pyre would still think of it as a protocol. To prevent the issue, keep track of protocols as a set and remove protocols per handle in the `purge` function.

Moved the DependencyHandler since we're relying on it to register protocols now.

Caught by the consistency checker.

Reviewed By: dkgi

Differential Revision: D14736415

fbshipit-source-id: 7beb9cf6869515b43c9f5c435bb854cd85da84f5