Prefer first overload
Summary:
This isn't actually specified yet, but it looks like the behavior of picking the first overload that matches is going to be specified soon (https://github.com/python/typing/issues/253).
For some reason we thought we needed to reverse the list to make this work, but evidently it wasn't reversed, so we were reversing it.

Reviewed By: sinancepel

Differential Revision: D14717465

fbshipit-source-id: 646dd01ff50b709a3ed55a9905fe39611d935286