Taking advantage of improved solve_less_or_equal in order to implement sound incomplete type handling.
Summary:
The ultimate objective of all of this solve_less_or_equal work was to solve the problem of our unsound and incomplete post-determination process.  The existing system was based on replacing unbound type variables with bottom, then replacing those bottoms with free type variables down the line to let subsequent calls narrow that down.  That strategy was fundamentally unsound due to the ability to have access to that incomplete value through a name, leading to potential contained type conflicts.  Additionally, the existing system handled cases like `x: List[int] = []` by special casing that bottom matches against any value regardless of variance.  Unfortunately this special casing only worked one layer deep, leading to issues with assignments like `x: Dict[str, List[int]] = {"A": []}`.

The new approach is to allow unbound type variables to escape signature selection, just marked as being escaped.
The rules for these escaped types is that they are only allowed to be returned, passed in as a parameter, or assigned to a variable with an explicit type annotation.  They are not allowed to be have an attribute accessed, be called, or assigned to a variable without an explicit annotation.  In any of those cases, as expected by PEP484, the escaped variables are converted to Anys.

The only problem I see remaining is issues like `x: Tuple[List[int], List[str]] = ([], [])` as this is trying to solve `Tuple[List[_T], List[_T]] ≤ Tuple[List[int], List[str]]`.  I will fix this in a follow up diff where we make namespace a global fresh variable value incremented at the process level.

Reviewed By: dkgi

Differential Revision: D14677823

fbshipit-source-id: fe2633b4571222a7fce89b02016671aa3f1f0e52