Remove define body from `Error.t`
Summary:
This diff is a direct followup to D14737822. See the summary in that diff for backgrounds and motivations. Most of the work is just about switching from `Define.t` to `Define.signature` in `Error.t` and making sure the OCaml type checker is happy.

The only troubling child is `IncompatibleReturnType` error. During error filtering we want to know whether the incompatibility is caused by an unimplemented function body. This check requires us to examine `define.body`, which is exactly what we are trying to remove. Fortunately, the check can be shifted from the error filtering phase to the type checking phase where the `IncompatibleReturnType` error is actually created.

Reviewed By: shannonzhu

Differential Revision: D14752250

fbshipit-source-id: 06a750356a163e196f2798365a3e347106f0c0ac