Implementing IntVar
Summary:
This diff implements the IntVar concept as described by Ivan Levkivskyi at the 3/12/19 Typing meetup.
This could not be implemented as a normal upper bounded int as upper bounds are inclusive, and we only want Literal integers in this variable, and not normal ints or other int subclasses.  In effect this is a `TypeVar("T", bound = Literal[..., -1, 0, 1, 2, ...])`.
As described in the slides, this feature could be extended to support arithmetic on these type variables, i.e `def concat(left: SizedList[T1], right: SizedList[T2]) -> T1 + T2`, but this will be deferred.

Reviewed By: sinancepel

Differential Revision: D14736704

fbshipit-source-id: 982863e86387e7d0355ba675d562ec039bb40aab