Use reference in protocol queries
Summary: Migrated the queries that are not types -- note, types cannot be represented as references. This is actually just fine, because the input format is always "AST" and types and references are two backend representations of that AST that we keep around for analysis. In many ways Reference.t is just a parallel to Type.t.

Reviewed By: dkgi

Differential Revision: D14762913

fbshipit-source-id: b938a5ea02a9b3eb1485b64802b62ce49cd27478