account for ambiguity in model definitions of class/instance attributes
Summary:
```
Class.ATTRIBUTE: TaintSource[...]
```
currently can be either denote that the attribute `ATTRIBUTE` of instances of
class `Class` are sources or that the class attribute `ATTRIBUTE` is a source.

Accounting for both in the model lookup assuming that at runtime there is no
ambiguity possible: fields cannot be both class and instance attributes.

Reviewed By: fahndrich

Differential Revision: D14766676

fbshipit-source-id: a9802b27481b4ada2b6698ad36ce0c72981abc32