add instead of writing through when adding values in environment building
Summary:
We were using `SharedMemory.write_through` at multiple places throughout to codebase - turns out that this doesn't play well with the local changes transactions we make, in which the .write_through you do ignores the local changes completely, and loses the behavior of latest writer wins.

This manifested itself as a bug in class definition refinement, where in incremental runs on the server (where we already had a class written), the first writer would win and class definitions wouldn't get refined. This meant that the method resolution order, etc., was broken.

Basic discrepancy found by the consistency checker.

Testing this was a bit hard - I added a simple populate test, hope that's enough for folks.

Reviewed By: dkgi

Differential Revision: D14766733

fbshipit-source-id: 788d75c4ac2f81572dfa9674c9a9dc1c962bd5a3