Using resolve_mutable_literal in mismatch_with_any check
Summary:
During type checking, we treat Any like Top, i.e. not a subtype of anything.  We then filter out mismatch errors that failed `given ≤ expected`, but can pass `given is_consistent_with expected`.  is_consistent_with is defined to be the same as ≤ except that Any is_consistent_with anything and vice versa.

Separately, we handle the issue of returning [1] when expecting list[float] by special casing situations where the given expression is a literal dict list or set, and try to weaken our inference to make the match work.

The problem is we don't have access (currently) to the expression, so can't do the weakening in the filtering step, and thus can't tell the difference between a named List[Any] and a literal `[function_returns_any()]`.

Reviewed By: dkgi

Differential Revision: D14726245

fbshipit-source-id: 7d29495b7ca7a83d2cfd16127202bc11be6894ce