split up class definition from class metadata
Summary:
The "class definition" name in the environment handler was a misnomer - the definition also held metadata that was computed later on, and had two states that are not obvious from an API view. Separate the metadata from the class definition itself; we usually only need one or the other, and bundling the data is cumbersome both from a performance and user perspective.

This also makes the consistency checker higher signal, since the discrepancies that had to do with class definitions were almost always class metadata discrepancies.

Reviewed By: dkgi

Differential Revision: D14779154

fbshipit-source-id: d62135672b67bdf74f3e79ac0025f72bcb05cfd7