properly prefer stubs over sources in search path subdirectories
Summary:
Our filtering logic for files was incorrect in the case where we wanted to add the subdirectory of a directory in the search path - we'd try to reconstruct a handle against the search root, not the handle root, meaning that if our search path was { root: "/root/", subdirectory: "analyze" }, /root/analyze/stub.pyi would have a handle `anaylze/stub.pyi` and get reconstructed as `/root/analyze/analyze/stub.pyi`, which is not what we were hoping for. Change this search so that we pass in the SearchPath structure and get the actual root/search root based on what we need.

The sources were getting registered over stubs in these cases because the find_sources function didn't share the same bug.

Reviewed By: grievejia

Differential Revision: D14782047

fbshipit-source-id: 7407eaf4efa1b03ec5a82fab44f75c78f9bab591