Use references in import module
Summary:
Held off on this for a while thinking it would make sense for the components of Import to be converted to Attribute AST objects, since the name does go through access fold for errors and it does map to real locations in the code.

Decided that we can migrate this over to use those AST objects after the refactor, and for simplicity now I'll use references. We're not losing any information, since locations were not stored on the old Access.t, so functionally this is identical. It would just make migration more complicated as this is the only AST object that references back to AST Accesses, but open to pushback on this.

Reviewed By: dkgi

Differential Revision: D14784440

fbshipit-source-id: a1ebac4927c87aeb5c713b6d2321d80344a7f659