Avoid invoking `Error.show` in `Error.join_at_define`
Summary:
Profiling results suggest that `Error.join_at_define` can be a real performance bottleneck on large files with many errors, and the the culprit is the `Error.show` within it.

This diff rewrote `Errror.join_at_define` such that `Error.show` is not called. Instead, we separate error deduplication into a standalone function `Error.deduplicate`. Both `joint_at_define` and `deduplicate` relies on the ppx-generated `hash` function directly, obviating the need of `show` and the associated cost of string allocations.

Reviewed By: dkgi

Differential Revision: D14719529

fbshipit-source-id: 147f5f38119383f0fb1b93942596a6df8f7302ac