Use references in preprocessing qualification logic
Summary:
This will require more work when remaining AST accesses are migrated to attributes.

Might seem counterintuitive to convert AST accesses to references when storing them in alias lookup, but this is actually an important distinction - we *don't* want raw AST to be stored as a lookup and then tacked on as a prefix to some other AST later. When migrating this to attribute AST objects, we'll just have to be careful about how we convert references back to AST (the locations will all be set to Location.Any) and maybe set those alias locations to be the location of the first element of the attribute we're prefixing it to. Currently none of this is an issue because there are locations nowhere.

Reviewed By: dkgi

Differential Revision: D14795317

fbshipit-source-id: a41e9cf90896535cefd8aff5b02355bdc1f7c94b