fix removal of dependents
Summary:
Turns out that our invalidation of dependents was broken in the shared memory handler - we were never removing dependents before re-adding, meaning that they never actually got invalidated. Caught by the consistency checker.

Add a test that ensures the purging behavior works as expected in memory, and restructure how we invalidate dependents a bit to make it easier to read.

Reviewed By: dkgi

Differential Revision: D14840685

fbshipit-source-id: 3115f810ded119f3ead3bfe607e38cb692d040b9