more precise search path for subdirectories
Summary:
In some edge cases, we can have overlapping subdirectories imported in a search path which should result in different handle structures, but end up matching the handle of the first one found.

For instance, if our search path is ["root$a"; "root/subdirectory$b"], the previous iteration of the search would search over the path ["root"; "root/subdirectory"], effectively meaning that the second element was never used. This change makes it so that the search path is ["root/a"; "root/subdirectory/b"], and handles are fixed up later.

Reviewed By: grievejia

Differential Revision: D14856530

fbshipit-source-id: edc4358e7cdcb5b56347e51896475f514417d566