Find buck-out directories in non-fbcode repositories
Summary:
howdidthishappen
We were throwing away the values of the map from given target -> buck link tree destination, and then re-constructing this target manually in the find_built_source_directories logic.

This is silly, and also it breaks in non-fbcode directories where the destination does not follow the buck-out/gen formula. Refactor this a bit to keep those values around and use them directly to check whether a target has already been built or not.

Reviewed By: DSouzaM

Differential Revision: D14833263

fbshipit-source-id: fddef404996da47ff3a86e93cac52878c0f32db0