Traverse Call in transform logic
Summary: Need this to be implemented in transform for the rest of preprocessing logic to be migrated, but it's a little tricky because we don't actually want to branch into nested expressions when we are converting new AST objects back to old accesses. The conversion logic in Expression.ml assumes that when it sees a new AST object, the nested objects are also new AST objects. The recursion is done there anyway, so traversing all the nested expressions here is wasted work.

Reviewed By: grievejia

Differential Revision: D14861169

fbshipit-source-id: 48f6f6575dcded71326c3b052d5433b13544b665