handle type order crashes while filtering errors
Summary: Now that we utilize heavyweight type order machinery for filtering errors, we can have cases where the given return annotation is an invalid type. The type order checks will throw an Untracked exception when this happens - default to surfacing the error to avoid the type checker crashing mid-flight, while not hiding information from the user.

Reviewed By: dkgi

Differential Revision: D14861197

fbshipit-source-id: 46567a83266e4e1d37ae49c69643c26e23ec2567