add BuckBuildStrategy abstraction to make buck algorithm configurable
Summary: On creation of the SharedAnalysisDirectory, we can pass a build strategy, which takes a list of targets as input and returns a list of output directories. This moves the decision about what build strategy to use to one place, and makes it easy to integrate the new buck builder.

Reviewed By: sinancepel

Differential Revision: D14859285

fbshipit-source-id: 1bd55c4232e57648470eedd03316379b20a88e6d