fix detection of built link trees
Summary:
Pyre is having trouble finding link trees built by Buck because the glob pattern doesn't have a `/` separator between the directory and the "*#*link-tree" part.
Also added the `buck_root` prefix so that the path is absolute, and added the target name so that the link tree we find is for the specific target (and not another one in the same build file).

Reviewed By: shannonzhu

Differential Revision: D14889878

fbshipit-source-id: 1e399dcfce3ec521b9c1642c55a80c56c2b51ab6