Refactor placeholder tuples
Summary: The old way of handling placeholder tuples caused us to create extra trace frame objects that for the most part copied over information from another trace frame. This refactor allows us to simply use the same trace frame object (since we're working with the same information), but treat it differently based on whether or not the placeholder flag is set.

Reviewed By: zlandau

Differential Revision: D14881234

fbshipit-source-id: a0f18d62b26f99986924d78afeedb6ef5d6b50cb