Prevent leaking namespaced variables from simulated signature select
Summary: Without this check, we can indirectly leak the internal variables of a simulated signature select through, unescaped.  We don't want to apply this restriction universally in extract_partial_solution, as this is intended behavior when doing the extraction at the end of the real signature select, as we don't use the remaining constraints in that case.

Reviewed By: sinancepel

Differential Revision: D14881738

fbshipit-source-id: 44057fc08f8d17954d07fe3cb0343957d3725233