Add logic to filter out duplicate locations
Summary:
While testing - I noticed that sometimes the server returns more than one type in the same location. This was causing some weird issues when printing, and for the most part the types returned in the same location were the same coverage level.

This adds logic so we can use set to filter them out.

Reviewed By: dkgi

Differential Revision: D14919385

fbshipit-source-id: 954efbbb640699e07081b282b46adb91197aa86e