Free variable ≤ free variable constraints must have either option, not necessarily both
Summary:
Because our constraint solve wants to be able to sequentially solve independent constraints, it was not set up to be able to solve statements of the form V1≤V2, as resolving each one depended on the other.  Rather than overcomplicating the solve to do actual elimination etc., we can get much of that functionality by recognizing that this is a disjunction of constraints rather than a conjunction, as a constraint on V1 to that effect implies the constraint on V2 and vice versa.  We can thus take advantage of our existing system for considering disjoint constraint sets to consider both options here, which allows us to handle the previously unhandlable case of passing a generic identity function to `map`.

TL;DR:
Circular constraints (A constrained by B AND B constrained by A) make the solver sad.
We can enforce that one variable be less than another one without a circular constraint (A ≤ B OR B ≤ A).

Reviewed By: sinancepel

Differential Revision: D14882351

fbshipit-source-id: 1b7ae94c677cf40a6ef96845dcb4154400e19e7c