Use real fresh variables
Summary:
In order to accept cases like
```
 def foo() -> typing.Tuple[typing.List[int], typing.List[str]]:
   return [], []
```
we need to be able to distinguish between two different versions of the same type variable.  Doing this without process level globals would require deconflicting every time we put two types together, which would be exceedingly annoying, fragile and slow.
We have to reset the index on every statement in order to prevent issues with the fixpoint, as otherwise the types inside of the errors generated from these escaped variables would not match and create an infinite loop.  This is fine because escaped variables and simulated type variables are not permitted to escape a statement anyway.

Reviewed By: sinancepel

Differential Revision: D14792782

fbshipit-source-id: 89f558ec8be82b6b17a997ee5d0bcd971d6bf680