keep file monitor alive indefinitely
Summary: Previously the monitor had a 1-hour timeout, so it would disconnect if it did not receive any watchman updates after an hour. A common pattern occurred: come back to a project after a break, edit files, then run `pyre incremental` to find that the file monitor missed all of the changes because it was stopped. Disables the timeout entirely, so it should stay alive indefinitely.

Reviewed By: sinancepel

Differential Revision: D14950883

fbshipit-source-id: 9172cb9df5c1f5112e50b63efc66a747184474d2