Remove all conversion to old accesses in testing
Summary:
Ensure nothing is borked after all this refactoring.

Reasoning: we should only ever be converting the AST we parse in tests back to old accesses if we are trying to parse the "expected" value. The "actual" value, which is calculated through real logic, should never be converted. Sources used to build the environment should never be converted. This way, we can feel assured that we don't have passing tests just because we convert things to old ASTs, but at runtime, logic is actually breaking because we forgot to gate something.

To find use-cases in tests:
grep -R --exclude-dir={website,_build,.pyre,hack_parallel,docs} 'convert' .

Reviewed By: dkgi

Differential Revision: D14936513

fbshipit-source-id: 7de7a4ada6e51ddc262ff0c7942f7b0ce1070599