Cache expensive queries in `Resolution.t`
Summary:
Similar to what we did in D14816849. This time we cache queries for `Resolution.t`. Note that we don't do this for all queries --- only the ones that I think are worth caching. We generally do not need to worry about invalidating the cache during the fixpoint computation as long as the queries themselves do not rely on the moving part of `Resolution.t` (namely `annotations` and `type_variables`), which should be the case right now.

In principle, I think the best place to put these caches is some sort of global state that only persists during type checking (the same can be said for `Annotated.Class.Attribute.cache`, `Type.cache`, and `Module.cache`). In fact, most fields in `Resolution.t` should work this way, but for historical reasons we do not do that and it's also hard to change

Instead, we follow what was done for the other caches by utilizing a global module. Invalidation becomes trickier, but should still be managable.

Reviewed By: dkgi

Differential Revision: D14955873

fbshipit-source-id: bd952a6e291e9768ef2a6ebf968ac9fe90316afa