Make new AST the default for all flags
Summary:
Setting the default for functions that deal with conversions between ASTs to default to using the AST unless the convert flag is specified as true.

Having half/half is confusing and extremely bug-prone; it's hard to spot some `Type.expression` somewhere in logic that's being refactored that we forgot to tack on a `~convert:false`, and then forget to migrate some part of the tests... results in bad combinations of tests maybe passing but logic not actually working at runtime.

Also as a side-bonus, progress on this entire refactor is a bit more visible now that we can track getting converts down to 0.

Reviewed By: dkgi

Differential Revision: D14962292

fbshipit-source-id: 72a9dc1c4f17c907006c0153dc0d00d4f722ab7d