Move instantiation into attributes from attribute
Summary:
You would expect calling `attributes` with an `instantiated` parameter would give you attributes with the generic parameters instantiated. Instead, that functionality is only done by `attribute`.  This is fine because (currently) attribute is the only thing that calls attributes with a `instantiated` parameter, but this is simpler and actually faster b/c now that instantiation is cached.

This diff also changes `attributes` to take an optional type rather than an optional optional type.  Not sure why that was done.

Reviewed By: grievejia

Differential Revision: D15000599

fbshipit-source-id: df684081ce115daa4a19d21464502687216a4b80