Don't abort parsing if Buck build file has assignment statements
Summary: We were asserting that every statement in the body of a Buck build file was an Expr, but sometimes users define variables in their build files. Until we properly support these assignments, we should ignore these statements. If a variable is used in a target we attempt to parse, we will still fail (since we don't know how to interpret variables), so this shouldn't hide any errors.

Reviewed By: dkgi

Differential Revision: D15019894

fbshipit-source-id: 7e9a3f1e9733c7b708ae5d0efac491e60213eaea