make apply_decorators return a overload
Summary:
The majority of this diff is dealing with the fallout of making apply_decorators return an overload, which is what it should really be doing - creating phantom defines was a hack that we used for a while; decorators deal with types.

The fallout is that we rely on apply_decorators for different use cases, such as Attribute.create and verifying the parameters in backwards analysis. The key issue is that we can't have apply_decorators return a callable; a callable constitutes of many overloads, and a decorator only works on a single overload. To accomodate this, reworked Callable.create to take in a list of overloads, together with the information of whether they're a implementation or overload (yay, python).

No functional change is intended; not sure I handled the taint analysis changes in the best way possible.

Reviewed By: dkgi

Differential Revision: D14984021

fbshipit-source-id: 1d2fa604096a732057f7114a67adaf90301c51f6