store undecorated functions
Summary:
This is the final step we need in order to get a limited implementation of decorator support - we need access to function types when registering functions, which obviously doesn't exist. The longer term solution is to respect dependency order when adding files to the environment in order to support decorators, but this lays the groundwork to handle undecorated decorators

I separated this change from the implementation of decorators to make code review easier.

Reviewed By: dkgi

Differential Revision: D14986633

fbshipit-source-id: ae86ef560dcf4014f1cc195cd53f003b062a7c1d