support user decorators
Summary:
Supports a limited form of user-specified decorators by using pre-registered functions when adding globals to the environment.

There are a couple limitations.

1. This explicitly does not support decorated decorators. I don't suspect this will be a huge problem in practice.

2. We've incurred a memory overhead because we don't respect the import order when building the environment. This would be possible without adding a function registration phase if we implement dependency-aware parallel registration (which would also incidentally help with aliasing).

3. We only support decorators that map callables to callables, and explicitly do not support decorators such as property. The reason for this is the interaction with overload. We can only construct a function from overloads, so it's not clear what it means to have a `property` `overload`, but this is something that the type system allows. It was more convenient for decorators to map callables to callables rather than having to deal with arbitrary type -> type, although it is obviously more limited.

Note: I had to change how behavioral subtyping works with anonymous parameters to support this diff.

In behavioral subtyping, the parameters of a overriding function must be more permissive than the overridden one. One particular case of this we hadn't run into until decorators was that a anonymous name is less permissive than a named parameter (since you can only call it positionally). So, it is consistent to allow overriding a anonymous parameter with a named one.

The strategy taken to achieve this was naming anonymous overridden parameters with the parameter at the corresponding position.

The reason we need this is the following:

```
T = TypeVar("T")

def decorate(f: Callable[['C', T], None]) -> Callable[['C', T], None]:
  ...

class C:
  decorate
  def f(self, x: int) -> None: # registered type is Callable[[C, int], None]
    pass

class D(C):
  # Since the type system disallows calling f with `x=1`, this is ok.
  def f(self, y: int) -> None:
    pass
```

Reviewed By: dkgi

Differential Revision: D15011034

fbshipit-source-id: 3ae8de12a9e937e3d7e31b0b79809959ec1cf2a5