hardcode common server inconsistency issue
Summary:
It's hard to write a test for this as it comes up in a server inconsistency.

The concrete issue that goes wrong is for IntEnums.

We'd like `__getitem__` for a `IntEnum` to resolve to `EnumMeta.__getitem__`, which is true since `IntEnum` has `int` and `Enum` as bases (`Enum` has `EnumMeta` as a metaclass). Now, after protocols are inferred, we hit the known issue that these inferred protocols are added to the method resolution order (ClassMetadata.superclasses is where they are stored). The problem in this case is that `int` implements the `SupportsAbs` generic protocol. The `SupportAbs` generic protocol has `Generic` has a superclass, so we end up using that `typing.Generic.__getitem__` instead of going through the meta hierarchy.

Reviewed By: dkgi

Differential Revision: D15016510

fbshipit-source-id: 14b89484f202081975a1598f6da8f07165e82294