Extract platform selection to separate module
Summary: The platform for a given project is necessary to build a lot of rules completely, so we extract this "inference" logic out to a separate module that can be used from various locations in the code. Right now, we just assume a particular platform, but in the future this abstraction will make it easier to do more intelligent inference.

Reviewed By: sinancepel

Differential Revision: D15031986

fbshipit-source-id: 18e2bf500ab88e75db079f8439473a444854474e