Supporting recursive protocols in principle
Summary:
This diff introduces a mechanism to handle all of our protocol needs, instantiate_protocol_parameters.
This function gives *a* solution [1] for the type variables of a protocol, such that the given candidate is a subtype, or None, if no such solution exists.
This is different than the existing implementation for a number of ways:
1. It supports recursive protocols by way of the "pseudoedge" system, which accumulates assumptions while recursing in order to avoid infinite loops.  This is also an easy point for caching.
2. It naturally supports attribute protocols.
3. It plays nice with Anys, as during filtering, the analysis correctly can now learn the edge differently, unlike the static solution

However, it is currently gated from production through the `is_protocol` function, which currently is always false during type checking runs.  The follow up diffs then enable it alongside the existing solution, then remove the existing solution.

[1] It only returns one possible solution, but in general this shouldn't be a problem, as long as the type variables associated with the protocol have the most permissive variance possible, in which case it handles things equivalent to the ideal case.  It should also be straightforward to complain or correct in cases where the variance is invariant when it could be one or the other.

Reviewed By: shannonzhu

Differential Revision: D14993295

fbshipit-source-id: 5fd8f3bda33173a8123c0a34b1dbb59ff704e963