Move `AnnotatedAttribute.Cache` back into `AnnotatedClass`
Summary: I feel bad about reverting mrkmndz's change, but I think it's not a good idea to expose the type of the cache in the mli file: It makes changing the key/value type of this cache much less hassle-free than it should be. Not exposing the cache type is also consistent with other existing caches in the codebase.

Reviewed By: sinancepel

Differential Revision: D15036366

fbshipit-source-id: c7c8fa7e51fc646c7e825ceedaf7dae41713c92e