Use a lookup table rather than a list for `AnnotatedClass.attributes`
Summary:
We solve the same problem as described in D14985868 here, except using the alternative approach mentioned there. I plan to abandon the original approach once this diff is accepted.

The basic idea here is to use a hashtable to hold the result of `AnnotatedClass.attributes`. This has the benefit of: (1) `AnnotatedClass.attriubute` can enjoy an `O(1)` lookup time once the underlying `attributes` gets computed. (2) no fields with duplicated names get generated, which saves some memory. (3) We optimized the asymptotic complexity of attribute list construction from `O(n^2)` to `O(n)`. (4) We can get rid of the ugly duplicated cache key presented in D14985868.

To make sure the change is non-invasive, we preserve the existing APIs of `AnnotatedClass`. Instead, we simply redirect the original `attributes` call to the new hash table implementation, and then convert the table back to a list (this conversion should not increase the asymptotic complexity assuming that the clients of `attributes` performs at least `O(n)` computation, which should be the case now). One issue there is that some client of `AnnotatedClass.attributes` (i.e. namedtuple resolution) implicitly depend on the order of the attributes in the list, so we cannot just replace the original attribute list with a single hash table as the latter preserves no order information. The solution is simple: In addition to the hash table, we also need to maintain a list of names in their reverse-insertion order, and rely on it to do the table-to-list conversion.

Reviewed By: sinancepel

Differential Revision: D15040858

fbshipit-source-id: 8fa291a227202cc9e635695e6b222a6c5b64c6e7