Optimize `Module.aliased_export` query
Summary:
Previously, we compute `aliased_export` as a `Map`, then immediately store it as an associative list. At query time, we convert this list back to `Map` again and then do the search, which is rather pointless since the conversion itself is `O(n)` so we gain nothing by using a `Map` for lookup.

We can just store the `Map` in the first place, and avoid the back-and-forth conversions. This also optimize `Module.aliased_export` query from `O(n)` to `O(log n)`.

Reviewed By: sinancepel

Differential Revision: D15067685

fbshipit-source-id: bd17a392c747e2b5c5a172c5c232aa2488dac43d