Optimize `Str` usage in AST-related functions
Summary:
`Identifier.sanitize`'s logic can be written in a more performant way, without using `Str`. It deserves a rewrite due to its hotness.

In `Source.parse`, we repeatedly compile the same regex again and again. Some of these usages can indeed be replaced by a hand-crafted function, but the gain may not be worth the effort.

Reviewed By: sinancepel

Differential Revision: D15083299

fbshipit-source-id: 6407e1c13b199dc98dc41e13f8b8403466ffeaf7