Using basic type partitioning for more principled isinstance checking
Summary:
This diff extends the ideas of D14545445 to apply to more than the special case of lists, and to work in the positive case as well.
The average case we're trying to support here looks like this:
```
x: Union[int, Set[int]]
if isinstance(x, set): # the runtime disallows using a parametric type here, and pyre will correctly interpret it as Set[Any]
   # x should be known to be List[int]
else:
   # x should be known to be int
```
In theory, the positive case should be inferred to be `Intersection[previous type, asserted type]` and the negative case to be `Difference[previous type, asserted type]`.
Without supporting these types first class, we can approximate them by opening up the unions  (and implicit unions like Optional), and doing element-wise comparisons with the boundary.
In order to support types with Anys, we need to use `is_consistent_with` rather than normal `less_or_equal`.

Reviewed By: sinancepel

Differential Revision: D15060401

fbshipit-source-id: 656b63a51b62d07e21f2c1b8fc700ca15b94a739